Revert IT tests
